requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
huey==2.3.2
idna==3.2
Markdown==3.3.4
psycopg2-binary==2.9.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
